Updates: Corrected the endpoints authentication and error handling

+++ src/userproject/userproject.controller.ts
@@ -12,6 +12,7 @@ import {
   UseGuards,
   NotFoundException,
   ForbiddenException,
+  HttpException,
 } from '@nestjs/common';
 import { UserprojectService } from './userproject.service';
 import { CreateUserprojectDto } from './dto/create-userproject.dto';
@@ -22,12 +23,14 @@ import { AuthGuard } from 'src/auth/Guards/auth.guard';
 import { AdminProjectGuard } from 'src/auth/Guards/adminProject.guard';
 import { User } from 'src/users/entities/user.entity';
 import { ProjectService } from 'src/project/project.service';
+import { UsersService } from 'src/users/users.service';
 
 @Controller('userproject')
 export class UserprojectController {
   constructor(
     private readonly userprojectService: UserprojectService,
     private readonly projectService: ProjectService,
+    private readonly usersService: UsersService,
   ) {}
 
   @UseGuards(AuthGuard, AdminProjectGuard)
@@ -38,6 +41,20 @@ export class UserprojectController {
     @Res() res: Response,
   ) {
     try {
+      const { project_id, user_id } = createUserprojectDto;
+
+      const [project, user] = await Promise.all([
+        this.projectService.findOne(project_id),
+        this.usersService.findOne(user_id),
+      ]);
+
+      if (user.role !== 'employee') {
+        throw new ForbiddenException('Only employess can be added');
+      }
+      if (req['user'].id !== project.pm_id.id) {
+        throw new ForbiddenException('Access Denied');
+      }
+
       const userproject =
         await this.userprojectService.create(createUserprojectDto);
       return sendResponse(
@@ -59,6 +76,14 @@ export class UserprojectController {
     @Res() res: Response,
   ) {
     try {
+      const project = await this.projectService.findOne(
+        userProjectData.project_id,
+      );
+
+      if (req['user'].id !== project.pm_id.id) {
+        throw new ForbiddenException('Access Denied');
+      }
+
       await this.userprojectService.removeUserFromProject(userProjectData);
       sendResponse(
         res,
@@ -68,34 +93,11 @@ export class UserprojectController {
         null,
       );
     } catch (error) {
-      throw new BadRequestException(error.message);
+      throw new HttpException(error.message, error.status||httpStatusCodes['Bad Request'])
     }
   }
 
-  @Get()
-  findAll() {
-    return this.userprojectService.findAll();
-  }
-
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.userprojectService.findOne(+id);
-  }
-
-  @Patch(':id')
-  update(
-    @Param('id') id: string,
-    @Body() updateUserprojectDto: UpdateUserprojectDto,
-  ) {
-    return this.userprojectService.update(+id, updateUserprojectDto);
-  }
-
-  @Delete(':id')
-  async remove(@Param('id') id: string) {
-    return this.userprojectService.remove(+id);
-  }
-
-  @UseGuards(AuthGuard, AdminProjectGuard)
+  @UseGuards(AuthGuard)
   @Get('/users/:projectId')
   async getUserbyProjectId(
     @Param('projectId') projectId: number,
@@ -107,6 +109,7 @@ export class UserprojectController {
 
       const users =
         await this.userprojectService.getUsersFromProject(projectId);
+
       if (!users) {
         throw new NotFoundException('Users does not exists for the project');
       }
@@ -121,15 +124,16 @@ export class UserprojectController {
             "You can not access other project's members list",
           );
         }
-        if (user?.role === 'employee') {
-          const data = users.filter((u) => {
-            return u.user_detail.user_id === user.id;
-          });
-          if (data.length === 0)
-            throw new ForbiddenException(
-              'You are not the part of this project',
-            );
-        }
+      }
+
+      if (user?.role === 'employee') {
+        const data = users.filter((u) => {
+          return u.user_detail.user_id === user.id;
+        });
+        console.log(data);
+
+        if (data.length === 0)
+          throw new ForbiddenException('You are not the part of this project');
       }
 
       return sendResponse(
@@ -140,7 +144,7 @@ export class UserprojectController {
         users,
       );
     } catch (error) {
-      throw new BadRequestException(error.message);
+        throw new HttpException(error.message, error.status||httpStatusCodes['Bad Request'])
     }
   }
 
@@ -163,7 +167,7 @@ export class UserprojectController {
         projects,
       );
     } catch (error) {
-      throw new BadRequestException(error.message);
+        throw new HttpException(error.message, error.status||httpStatusCodes['Bad Request'])
     }
   }
 }
